commands: move snapshot subcommands to subdirs
