+++ command/snapshot/inspect/snapshot_inspect_test.go
@@ -1,4 +1,4 @@
-package snapshotrestore
+package inspect
 
 import (
 	"io"
@@ -12,14 +12,14 @@ import (
 	"github.com/mitchellh/cli"
 )
 
-func TestSnapshotRestoreCommand_noTabs(t *testing.T) {
+func TestSnapshotInpectCommand_noTabs(t *testing.T) {
 	t.Parallel()
 	if strings.ContainsRune(New(cli.NewMockUi()).Help(), '\t') {
 		t.Fatal("usage has tabs")
 	}
 }
 
-func TestSnapshotRestoreCommand_Validation(t *testing.T) {
+func TestSnapshotInspectCommand_Validation(t *testing.T) {
 	t.Parallel()
 	ui := cli.NewMockUi()
 	c := New(ui)
@@ -59,24 +59,18 @@ func TestSnapshotRestoreCommand_Validation(t *testing.T) {
 	}
 }
 
-func TestSnapshotRestoreCommand_Run(t *testing.T) {
+func TestSnapshotInspectCommand_Run(t *testing.T) {
 	t.Parallel()
 	a := agent.NewTestAgent(t.Name(), ``)
 	defer a.Shutdown()
 	client := a.Client()
 
-	ui := cli.NewMockUi()
-	c := New(ui)
-
 	dir := testutil.TempDir(t, "snapshot")
 	defer os.RemoveAll(dir)
 
 	file := path.Join(dir, "backup.tgz")
-	args := []string{
-		"-http-addr=" + a.HTTPAddr(),
-		file,
-	}
 
+	// Save a snapshot of the current Consul state
 	f, err := os.Create(file)
 	if err != nil {
 		t.Fatalf("err: %v", err)
@@ -95,8 +89,26 @@ func TestSnapshotRestoreCommand_Run(t *testing.T) {
 		t.Fatalf("err: %v", err)
 	}
 
+	// Inspect the snapshot
+	ui := cli.NewMockUi()
+	c := New(ui)
+	args := []string{file}
+
 	code := c.Run(args)
 	if code != 0 {
 		t.Fatalf("bad: %d. %#v", code, ui.ErrorWriter.String())
+	}
+
+	output := ui.OutputWriter.String()
+	for _, key := range []string{
+		"ID",
+		"Size",
+		"Index",
+		"Term",
+		"Version",
+	} {
+		if !strings.Contains(output, key) {
+			t.Fatalf("bad %#v, missing %q", output, key)
+		}
 	}
 }
