+++ command/snapshot/save/snapshot_save.go
@@ -1,4 +1,4 @@
-package snapshotsave
+package save
 
 import (
 	"flag"
